requirements.txt
fastapi
uvicorn
python-dotenv
requests
supabase
jinja2
starlette
langchain-huggingface
langchain-groq
langgraph
redis
langsmith
neo4j
qdrant-client
sentence-transformers
langchain-text-spitter